requirements.txt
kafka-python==2.0.2
pyspark==3.4.0
pymongo==4.6.0
pandas==2.0.3
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.1.0
python-dotenv==1.0.0
